Fix page deletion
