+++ frontend/packages/schema-editor/src/components/TreeView/SchemaItem.module.css
@@ -17,8 +17,3 @@
 .isRef:global(.Mui-selected) {
   border-style: dashed;
 }
-
-.confirmDeleteDialogTrigger {
-  float: right;
-  padding: 0 var(--fds-spacing-3);
-}
