+++ frontend/packages/schema-editor/src/components/TreeView/SchemaItem.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { TreeItem } from '@mui/lab';
 import { useDispatch, useSelector } from 'react-redux';
 import { setSelectedId } from '../../features/editor/schemaEditorSlice';
@@ -19,10 +19,6 @@ import type { DragItem } from './dnd-helpers';
 import { useDatamodelQuery } from '@altinn/schema-editor/hooks/queries';
 import { useDatamodelMutation } from '@altinn/schema-editor/hooks/mutations';
 import { getRefNodeSelector, selectedIdSelector } from '@altinn/schema-editor/selectors/schemaStateSelectors';
-import { AltinnConfirmDialog } from 'app-shared/components';
-import { useTranslation } from 'react-i18next';
-import { deleteNode } from '@altinn/schema-model';
-import { removeSelection } from '../../features/editor/schemaEditorSlice';
 
 export type SchemaItemProps = {
   selectedNode: UiSchemaNode;
@@ -43,10 +39,8 @@ export function SchemaItem({
   index,
 }: SchemaItemProps) {
   const dispatch = useDispatch();
-  const { t } = useTranslation();
   const { data } = useDatamodelQuery();
   const { mutate } = useDatamodelMutation();
-  const [isConfirmDeleteDialogOpen, setIsConfirmDeleteDialogOpen] = useState<boolean>();
 
   const keyPrefix = isPropertiesView ? 'properties' : 'definitions';
 
@@ -71,11 +65,6 @@ export function SchemaItem({
   const onMove = (from: DragItem, to: DragItem) =>
     mutate(changeChildrenOrder(data, { pointerA: from.itemId, pointerB: to.itemId }));
 
-  const handleDeleteClick = () => {
-    mutate(deleteNode(data, selectedNode.pointer));
-    dispatch(removeSelection(selectedNode.pointer));
-  };
-
   return (
     <TreeItem
       nodeId={selectedNode.pointer}
@@ -83,30 +72,16 @@ export function SchemaItem({
       onClick={(e: any) => onLabelClick(e, selectedNode)}
       onFocusCapture={(e: any) => e.stopPropagation()}
       label={
-        <>
-          <DndItem index={index} itemId={selectedNode.pointer} containerId={base} onMove={onMove}>
-            <SchemaItemLabel
-              icon={getIconStr(refNode ?? selectedNode)}
-              key={`${selectedNode.pointer}-label`}
-              selectedNode={selectedNode}
-              refNode={refNode}
-              translate={translate}
-              hasReferredNodes={isPropertiesView ? false : referredNodes.length > 0}
-              openConfirmDeleteDialog={() => setIsConfirmDeleteDialogOpen(prevState => !prevState)}
-            />
-          </DndItem>
-          <AltinnConfirmDialog
-            open={isConfirmDeleteDialogOpen}
-            confirmText={t('schema_editor.datamodel_field_deletion_confirm')}
-            onConfirm={handleDeleteClick}
-            onClose={() => setIsConfirmDeleteDialogOpen(false)}
-            placement='bottom'
-            trigger={<div className={classes.confirmDeleteDialogTrigger} />}
-          >
-            <p>{t('schema_editor.datamodel_field_deletion_text')}</p>
-            <p>{t('schema_editor.datamodel_field_deletion_info')}</p>
-          </AltinnConfirmDialog>
-        </>
+        <DndItem index={index} itemId={selectedNode.pointer} containerId={base} onMove={onMove}>
+          <SchemaItemLabel
+            icon={getIconStr(refNode ?? selectedNode)}
+            key={`${selectedNode.pointer}-label`}
+            selectedNode={selectedNode}
+            refNode={refNode}
+            translate={translate}
+            hasReferredNodes={isPropertiesView ? false : referredNodes.length > 0}
+          />
+        </DndItem>
       }
     >
       {childNodesSorted.map((childNode: UiSchemaNode, childNodeIndex: number) => (
